+++ daemon/daemon_test.go
@@ -1,0 +1,10 @@
+package daemon
+
+import "testing"
+
+func TestDaemon(t *testing.T) {
+	// TODO: NewInstance
+	// TODO: LoadConfig
+	// TODO: Run
+	// TODO: All
+}