+++ dict/sync_test.go
@@ -1,0 +1,19 @@
+package dict
+
+import "testing"
+
+func TestSyncMap(t *testing.T) {
+	// TODO: NewSyncMap
+	// TODO: NewSyncMapFrom
+	// TODO: SyncMap.Get
+	// TODO: SyncMap.Set
+	// TODO: SyncMap.SetIf
+	// TODO: SyncMap.Delete
+	// TODO: SyncMap.DeleteKeys
+	// TODO: SyncMap.Clear
+	// TODO: SyncMap.Len
+	// TODO: SyncMap.Map
+	// TODO: SyncMap.ClearMap
+	// TODO: SyncMap.Keys
+	// TODO: SyncMap.Values
+}