+++ io/io_test.go
@@ -1,0 +1,10 @@
+package io
+
+import "testing"
+
+func TestIo(t *testing.T) {
+	// TODO: IsDir
+	// TODO: PathExists
+	// TODO: EnsurePathExists
+	// TODO: BaseFilename
+}