+++ io/read_test.go
@@ -1,0 +1,9 @@
+package io
+
+import "testing"
+
+func TestReadJSON(t *testing.T) {
+	// TODO: ReadJSON
+	// TODO: ReadJSONList
+	// TODO: ReadJSONMap
+}